+++ spec/cocaine/command_line_spec.rb
@@ -179,6 +179,15 @@ describe Cocaine::CommandLine do
     logger.entries.should include("\e[32mCommand\e[0m :: echo 'Logging!' 'bar'")
   end
 
+  it 'can still take something that does not respond to tty as a logger' do
+    output_buffer = StringIO.new
+    logger = ActiveSupport::BufferedLogger.new(output_buffer)
+    logger.should_not respond_to(:tty?)
+    Cocaine::CommandLine.new("echo", "'Logging!' :foo", :logger => logger).run(:foo => "bar")
+    output_buffer.rewind
+    output_buffer.read.should == "Command :: echo 'Logging!' 'bar'\n"
+  end
+
   it "logs the command to a supplied logger" do
     logger = FakeLogger.new
     Cocaine::CommandLine.new("echo", "'Logging!' :foo", :logger => logger).run(:foo => "bar")